+++ fluxdb/shardinject.go
@@ -19,6 +19,8 @@ import (
 	"encoding/gob"
 	"fmt"
 	"io"
+	"strconv"
+	"strings"
 
 	"github.com/dfuse-io/dstore"
 	"github.com/dfuse-io/shutter"
@@ -40,13 +42,56 @@ func NewShardInjector(shardsStore dstore.Store, db *FluxDB) *ShardInjector {
 	}
 }
 
+func parseFileName(filename string) (first, last uint32, err error) {
+	vals := strings.Split(filename, "-")
+	if len(vals) != 2 {
+		err = fmt.Errorf("cannot parse filename: %s", filename)
+		return
+	}
+
+	first64, err := strconv.ParseUint(vals[0], 10, 32)
+	if err != nil {
+		return 0, 0, err
+	}
+	first = uint32(first64)
+
+	last64, err := strconv.ParseUint(vals[1], 10, 32)
+	if err != nil {
+		return 0, 0, err
+	}
+	last = uint32(last64)
+
+	return
+}
+
 func (s *ShardInjector) Run() (err error) {
 	ctx, cancelInjector := context.WithCancel(context.Background())
-	s.Shutter.OnTerminating(func(_ error) {
+	s.OnTerminating(func(_ error) {
 		cancelInjector()
 	})
 
+	startAfter, err := s.db.getLastBlock(ctx)
+	if err != nil {
+		return err
+	}
+
+	zlog.Info("starting back shard injector", zap.Stringer("block", startAfter))
+	startAfterNum := uint32(startAfter.Num())
+
 	err = s.shardsStore.Walk(ctx, "", "", func(filename string) error {
+		fileFirst, fileLast, err := parseFileName(filename)
+		if err != nil {
+			return err
+		}
+
+		if fileFirst > startAfterNum+1 {
+			return fmt.Errorf("file %s starts at block %d, we were expecting to start right after %d, there is a hole in your block range files", filename, fileFirst, startAfter)
+		}
+		if fileLast <= startAfterNum {
+			zlog.Info("skipping shard file", zap.String("filename", filename), zap.Uint32("start_after", startAfterNum))
+			return nil
+		}
+
 		zlog.Info("processing shard file", zap.String("filename", filename))
 
 		reader, err := s.shardsStore.OpenObject(ctx, filename)
@@ -55,7 +100,7 @@ func (s *ShardInjector) Run() (err error) {
 		}
 		defer reader.Close()
 
-		requests, err := readWriteRequestsForBatch(reader)
+		requests, err := readWriteRequestsForBatch(reader, startAfterNum)
 		if err != nil {
 			return fmt.Errorf("unable to read all write requests in batch %q: %w", filename, err)
 		}
@@ -65,6 +110,7 @@ func (s *ShardInjector) Run() (err error) {
 			return fmt.Errorf("write batch %q: %w", filename, err)
 		}
 
+		startAfterNum = fileLast
 		return nil
 	})
 
@@ -75,7 +121,7 @@ func (s *ShardInjector) Run() (err error) {
 	return nil
 }
 
-func readWriteRequestsForBatch(reader io.Reader) ([]*WriteRequest, error) {
+func readWriteRequestsForBatch(reader io.Reader, startAfter uint32) ([]*WriteRequest, error) {
 	decoder := gob.NewDecoder(reader)
 
 	var requests []*WriteRequest
@@ -88,6 +134,10 @@ func readWriteRequestsForBatch(reader io.Reader) ([]*WriteRequest, error) {
 		if err != nil {
 			return nil, fmt.Errorf("unable to read WriteRequest: %w", err)
 		}
+		if req.BlockNum <= startAfter {
+			continue
+		}
 		requests = append(requests, req)
+
 	}
 }