+++ launcher/cli/apps.go
@@ -25,7 +25,8 @@ import (
 
 	blockmetaApp "github.com/dfuse-io/blockmeta/app/blockmeta"
 	"github.com/dfuse-io/bstream"
-	_ "github.com/dfuse-io/dauth/null" // register plugin
+	_ "github.com/dfuse-io/dauth/authenticator/null" // register authenticator plugin
+	_ "github.com/dfuse-io/dauth/ratelimiter/null"   // register ratelimiter plugin
 	abicodecApp "github.com/dfuse-io/dfuse-eosio/abicodec/app/abicodec"
 	"github.com/dfuse-io/dfuse-eosio/apiproxy"
 	dblockmeta "github.com/dfuse-io/dfuse-eosio/blockmeta"